+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Changelog for NeoFS Node
 - Invalid URI error text in `request create-container` CLI command (#3366)
 - Invalid object ID error text in `object delete` CLI command (#3366)
 - Old objects without a session token lifetime are not valid (#3373)
+- Node attributes are not updated in map after configuration/version change (#3363)
 
 ### Changed
 - Make `MaintenanceModeAllowed` network setting always allowed (#3360)